% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_arithmetic.R
\name{process_image_operation}
\alias{process_image_operation}
\title{Process image-image operation}
\usage{
process_image_operation(img1, img2, op)
}
\arguments{
\item{img1, img2}{Magick image objects}

\item{op}{Operation function}
}
\value{
New magick image object
}
\description{
Process image-image operation
}
