% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_arithmetic.R
\name{process_numeric_operation}
\alias{process_numeric_operation}
\title{Process image-numeric operation}
\usage{
process_numeric_operation(img, num, op)
}
\arguments{
\item{img}{Magick image object}

\item{num}{Numeric value}

\item{op}{Operation function}
}
\value{
New magick image object
}
\description{
Process image-numeric operation
}
